pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smartermcp"
version = "1.0.0"
description = "SmarterMCP OAuth platform with JWT authorization codes"
readme = "README.md"
authors = [{name = "SmarterCL", email = "tech@smarterbot.cl"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pyjwt[crypto]>=2.8.0",
    "redis>=5.0.1",
    "httpx>=0.25.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0",
    "flake8>=7.0",
]
prod = [
    "gunicorn>=22.0",
    "uvloop>=0.19.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"